[data_glove] merge the ip_acquire.py into glove_control_pub.py.
